requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==4.0.6
django-appconf==1.0.5
django-ckeditor==6.5.0
django-compressor==4.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-sass-processor==1.2.1
django-sort-order-field==1.2
entrypoints==0.3
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
mechanize==0.4.7
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
outcome==1.1.0
pexpect==4.6.0
Pillow==9.2.0
praw==7.5.0
prawcore==2.3.0
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
rcssmin==1.1.0
requests==2.22.0
requests-unixsocket==0.2.0
rjsmin==1.2.0
SecretStorage==2.3.1
selenium==4.1.3
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.1
soupsieve==2.3.1
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.26.9
wadllib==1.3.3
webencodings==0.5.1
websocket-client==1.3.1
wsproto==1.1.0
zipp==1.0.0
zope.interface==4.7.1
